Move policy view to shared scope
