+++ Tireless/Auth/Sign in/View/SignInView.swift
@@ -118,47 +118,6 @@ struct SignInView: View {
     }
 }
 
-struct PolicyView: View {
-    @State private var policyIsPresenting = false
-    @State private var termsIsPresenting = false
-    
-    var body: some View {
-        Text("By continuing you accept our")
-            .font(.system(.caption))
-            .foregroundColor(.secondary)
-        
-        HStack {
-            Button(action: {
-                self.policyIsPresenting.toggle()
-            }) {
-                Text("Privacy Policy")
-                    .font(.system(.caption).bold())
-                    .foregroundColor(.main)
-            }
-            .buttonStyle(GrowingButton())
-            .sheet(isPresented: $policyIsPresenting) {
-                WebView(url: URL(string: "https://pages.flycricket.io/tireless-1/privacy.html"))
-            }
-            
-            Text("and")
-                .font(.system(.caption))
-                .foregroundColor(.secondary)
-            
-            Button(action: {
-                self.termsIsPresenting.toggle()
-            }) {
-                Text("Terms of Use")
-                    .font(.system(.caption).bold())
-                    .foregroundColor(.main)
-            }
-            .buttonStyle(GrowingButton())
-            .sheet(isPresented: $termsIsPresenting) {
-                WebView(url: URL(string: "https://www.apple.com/legal/internet-services/itunes/dev/stdeula/"))
-            }
-        }
-    }
-}
-
 struct IconButton: View {
     enum ImageName {
         case custom(String)
